+++ spec/models/team_leaderboard_spec.rb
@@ -93,7 +93,6 @@ describe TeamLeaderboard do
     let!(:user1_swim_activity_2) { Fabricate(:swim_activity, user: user1, team: team, start_date: Time.now) }
     let!(:user2_activity_1) { Fabricate(:user_activity, user: user2, team: team, start_date: 2.months.ago) }
     let!(:another_activity) { Fabricate(:user_activity, user: Fabricate(:user, team: Fabricate(:team))) }
-    let!(:club_activity) { Fabricate(:club_activity, team: team) }
 
     TeamLeaderboard::MEASURABLE_VALUES.each do |metric|
       context metric do